refactor subdir structure
