+++ cli/beacon/internal/tokens/contextwindow.go
