+++ cli/beacon/internal/tokens/contextwindow_test.go
