+++ cli/beacon/internal/tokens/tokens_test.go
