+++ docs/cli/index.mdx
@@ -44,7 +44,7 @@ Top-level Agent Beacon CLI commands.
 | [`beacon rules`](/cli/rules) | Manage and author Beacon threat-detection rules. |
 | [`beacon scan`](/cli/scan) | Run threat-detection rules over local endpoint telemetry. |
 | [`beacon status`](/cli/endpoint-status) | Alias for `beacon endpoint status`. |
-| [`beacon token-usage`](/cli/token-usage) | Report token usage and attribution from the endpoint runtime log. |
+| [`beacon token-usage`](/cli/token-usage) | Report token usage and attribution from Beacon runtime logs. |
 | [`beacon version`](/cli/version) | Display the installed Agent Beacon CLI version. |
 
 ## beacon version
