+++ docs/cli/token-usage.mdx
@@ -1,11 +1,11 @@
 ---
 title: "beacon token-usage"
-description: "Report token usage and attribution from the Beacon endpoint runtime log"
+description: "Report token usage and attribution from Beacon runtime logs"
 ---
 
 ## Command Overview
 
-`beacon token-usage` reads Beacon endpoint events from the runtime JSONL log and summarizes token usage, cost, models, sessions, runs, harnesses, and repositories.
+`beacon token-usage` reads Beacon runtime events from the JSONL log and summarizes token usage, cost, models, sessions, runs, harnesses, and repositories.
 
 ```bash title="Command syntax"
 beacon token-usage [flags]
